Refactor schema entry classes into separate modules
