+++ src/schema/parser/attributeParser.ts
@@ -1,6 +1,8 @@
-import { SchemaAttribute, type SchemaEntryManager, type SchemaProperty } from '../entries'
 import { getElementTagName, type HedSchemaXMLCollection, type HedSchemaXMLObject } from '../xmlType'
 import { SchemaEntryParser } from './schemaEntryParser'
+import type SchemaEntryManager from '../entries/schemaEntryManager'
+import type SchemaProperty from '../entries/property'
+import SchemaAttribute from '../entries/attribute'
 
 export default class AttributeParser extends SchemaEntryParser<SchemaAttribute> {
   /**
