+++ src/schema/parser/unitClassParser.ts
@@ -1,12 +1,12 @@
-import {
-  type SchemaAttribute,
-  type SchemaEntryManager,
-  SchemaUnit,
-  SchemaUnitClass,
-  type SchemaUnitModifier,
-} from '../entries'
 import { getElementTagName, type HedSchemaXMLCollection, type DefinitionElement, HedSchemaXMLObject } from '../xmlType'
 import { SchemaDefinitionEntryParser } from './schemaEntryParser'
+
+import type SchemaUnitModifier from '../entries/unitModifier'
+import type SchemaAttribute from '../entries/attribute'
+import type SchemaEntryManager from '../entries/schemaEntryManager'
+import SchemaUnitClass from '../entries/unitClass'
+import SchemaUnit from '../entries/unit'
+
 import { IssueError } from '../../issues/issues'
 
 export default class UnitClassParser extends SchemaDefinitionEntryParser<SchemaUnitClass> {
